refactor: fix type errors for summary react nodes and for tif and type validation
