+++ libs/deal-ticket/src/components/deal-ticket/type-selector.tsx
@@ -1,13 +1,15 @@
-import { FormGroup, InputError } from '@vegaprotocol/ui-toolkit';
+import { FormGroup, InputError, Tooltip } from '@vegaprotocol/ui-toolkit';
 import { t } from '@vegaprotocol/react-helpers';
 import { Schema } from '@vegaprotocol/types';
 import { Toggle } from '@vegaprotocol/ui-toolkit';
-import type { ReactNode } from 'react';
+import { compileGridData, MarketDataGrid } from '../trading-mode-tooltip';
+import type { DealTicketMarketFragment } from './__generated__/DealTicket';
 
 interface TypeSelectorProps {
   value: Schema.OrderType;
   onSelect: (type: Schema.OrderType) => void;
-  errorMessage?: ReactNode;
+  market: DealTicketMarketFragment;
+  errorMessage?: string;
 }
 
 const toggles = [
@@ -18,8 +20,47 @@ const toggles = [
 export const TypeSelector = ({
   value,
   onSelect,
+  market,
   errorMessage,
 }: TypeSelectorProps) => {
+  const renderError = (errorType: string) => {
+    if (errorType === 'auction') {
+      return t('Only limit orders are permitted when market is in auction');
+    }
+
+    if (errorType === 'liquidity') {
+      return (
+        <span>
+          {t('This market is in auction until it reaches')}{' '}
+          <Tooltip
+            description={<MarketDataGrid grid={compileGridData(market)} />}
+          >
+            <span>{t('sufficient liquidity')}</span>
+          </Tooltip>
+          {'. '}
+          {t('Only limit orders are permitted when market is in auction')}
+        </span>
+      );
+    }
+
+    if (errorType === 'price') {
+      return (
+        <span>
+          {t('This market is in auction due to')}{' '}
+          <Tooltip
+            description={<MarketDataGrid grid={compileGridData(market)} />}
+          >
+            <span>{t('high price volatility')}</span>
+          </Tooltip>
+          {'. '}
+          {t('Only limit orders are permitted when market is in auction')}
+        </span>
+      );
+    }
+
+    return null;
+  };
+
   return (
     <FormGroup label={t('Order type')} labelFor="order-type">
       <Toggle
@@ -29,7 +70,7 @@ export const TypeSelector = ({
         checkedValue={value}
         onChange={(e) => onSelect(e.target.value as Schema.OrderType)}
       />
-      {errorMessage && <InputError>{errorMessage}</InputError>}
+      {errorMessage && <InputError>{renderError(errorMessage)}</InputError>}
     </FormGroup>
   );
 };
